Refactor handling of storage states: https://github.com/vitalets/playwright-bdd/issues/184
